adminController // dispatcher // writerControllers

+++ Controllers/AdminController.php
@@ -1,8 +1,7 @@
 <?php
 
-class AdminController extends AppController
+class AdminController extends RegisterController
 {
-
     protected function beforeRender()
     {
         //verif si c'est un admin
@@ -10,7 +9,7 @@ class AdminController extends AppController
 
             //si reçois add_users avec toutes les données nécessaires
             if (!empty($_POST['add_user'])) {
-                $this->add_user();
+                $this->checkDatas();
             }
 
             //si reçois modif_users avec tous les champs user sauf password
@@ -27,43 +26,55 @@ class AdminController extends AppController
             $this->_params['data'] = $this->_model->users();
 
             $this->_params['users']=[];
-            $i=0;
+
             while ($data = $this->_params['data']->fetch(PDO::FETCH_ASSOC)) {
                 array_push($this->_params['users'], $data);
             }
             array_shift($this->_params);
-        }else {
-          $this->redirect();
+        } else {
+            $this->redirect();
         }
     }
 
-    private function add_user()
+    protected function add_user()
     {
         $input=[];
-        foreach ($_POST['add_user'] as $key => $value) {
+        foreach ($_POST['add_user'] as $key => $value)
+        {
             $input[$key] = $this->secure_input($value);
         }
-        if ($input['password'] == $input['password_confirm'])
-          {
-            $input['password'] = password_hash($input['password'],PASSWORD_DEFAULT);
-            $this->_model->add_user($input);
-          }else {
-            $_SESSION['message'] = "<p class='error'> password not match </p>";
-          }
+
+        $this->_model->add_user($input);
+        return true;
     }
 
     private function modif_user()
     {
-        $input=[];
-        foreach ($_POST['add_user'] as $key => $value) {
-            $input[$key] = $this->secure_input($value);
+        $email = $_POST['modif_user']['email'];
+        $res = $this->_model->check_email($_POST['modif_user']['email']);
+        //var_dump($res);
+        if (!empty($res)) {
+            if ($res[0]['email'] == $email) {
+                $_SESSION['message'] = "This email is already taken";
+                return false;
+            }
+            $test_group = ['ADMIN','WRITER','USER'];
+
+            if (in_array($_POST['modif_user']['group'], $test_group)) {
+                foreach ($_POST['modif_user'] as $key => $value) {
+                    $this->_datasUser[$key] = $this->secure_input($value);
+                }
+                $this->_model->modif_user($this->_datasUser);
+            }
         }
-        $this->_model->modif_user($input);
     }
 
     private function delete_user()
     {
         $id = $_POST['delete_user'];
-        $this->_model->delete_user($id);
+        $check_admin = new log($id);
+        if ($check_admin != "ADMIN") {
+            $this->_model->delete_user($id);
+        }
     }
 }
